+++ src/handlers/userChangemakerPermissionsHandlers.ts
@@ -1,4 +1,8 @@
-import { createOrUpdateUserChangemakerPermission } from '../database';
+import {
+	assertUserChangemakerPermissionExists,
+	createOrUpdateUserChangemakerPermission,
+	removeUserChangemakerPermission,
+} from '../database';
 import {
 	isAuthContext,
 	isId,
@@ -13,6 +17,61 @@ import {
 	InputValidationError,
 } from '../errors';
 import type { Request, Response, NextFunction } from 'express';
+
+const deleteUserChangemakerPermission = (
+	req: Request,
+	res: Response,
+	next: NextFunction,
+): void => {
+	const { userKeycloakUserId, changemakerId, permission } = req.params;
+	if (!isKeycloakUserId(userKeycloakUserId)) {
+		next(
+			new InputValidationError(
+				'Invalid userKeycloakUserId parameter.',
+				isKeycloakUserId.errors ?? [],
+			),
+		);
+		return;
+	}
+	if (!isId(changemakerId)) {
+		next(
+			new InputValidationError(
+				'Invalid changemakerId parameter.',
+				isId.errors ?? [],
+			),
+		);
+		return;
+	}
+	if (!isPermission(permission)) {
+		next(
+			new InputValidationError(
+				'Invalid permission parameter.',
+				isPermission.errors ?? [],
+			),
+		);
+		return;
+	}
+
+	(async () => {
+		await assertUserChangemakerPermissionExists(
+			userKeycloakUserId,
+			changemakerId,
+			permission,
+		);
+		await removeUserChangemakerPermission(
+			userKeycloakUserId,
+			changemakerId,
+			permission,
+		);
+		res.status(204).contentType('application/json').send();
+	})().catch((error: unknown) => {
+		if (isTinyPgErrorWithQueryContext(error)) {
+			next(new DatabaseError('Error deleting item.', error));
+			return;
+		}
+		next(error);
+	});
+};
 
 const putUserChangemakerPermission = (
 	req: Request,
@@ -86,6 +145,7 @@ const putUserChangemakerPermission = (
 };
 
 const userChangemakerPermissionsHandlers = {
+	deleteUserChangemakerPermission,
 	putUserChangemakerPermission,
 };